fix: uses config file for connection.
